+++ business/goods/goods-api/src/main/java/study/daydayup/wolf/business/goods/api/dto/loan/LoanDTO.java
@@ -3,6 +3,8 @@ package study.daydayup.wolf.business.goods.api.dto.loan;
 import study.daydayup.wolf.business.goods.api.vo.Loan;
 import study.daydayup.wolf.framework.layer.api.DTO;
 
+import java.math.BigDecimal;
+
 /**
  * study.daydayup.wolf.business.goods.api.dto.loan
  *
@@ -13,7 +15,7 @@ public class LoanDTO extends Loan implements DTO {
     /**
      * 精度 0.0000%
      */
-    private Double handlingFeeRate;
-    private Double interest;
-    private Double penalty;
+    private BigDecimal handlingFeeRate;
+    private BigDecimal interest;
+    private BigDecimal penalty;
 }
